+++ SCrawler/API/OnlyFans/Declarations.vb
@@ -17,5 +17,23 @@ Namespace API.OnlyFans
             {{"files", "full", "url"}}
         }
         Friend Property Rules As DynamicRulesEnv
+        Friend Class CookieDateProvider : Implements ICustomProvider
+            Private ReadOnly DefaultProvider As IFormatProvider
+            Friend Sub New()
+                DefaultProvider = PersonalUtilities.Tools.Web.Cookies.CookieKeeper.DateProviderDefault
+            End Sub
+            Friend Function Convert(ByVal Value As Object, ByVal DestinationType As Type, ByVal Provider As IFormatProvider,
+                                    Optional ByVal NothingArg As Object = Nothing, Optional ByVal e As ErrorsDescriber = Nothing) As Object Implements ICustomProvider.Convert
+                If Not IsNothing(Value) AndAlso TypeOf Value Is String AndAlso Not CStr(Value).IsEmptyString Then
+                    Dim v$ = Value
+                    If v.Contains(",_") Then v = v.Substring(0, v.IndexOf(",_"))
+                    Value = v
+                End If
+                Return AConvert(Of Date)(Value, DefaultProvider, Nothing)
+            End Function
+            Private Function GetFormat(ByVal FormatType As Type) As Object Implements IFormatProvider.GetFormat
+                Throw New NotImplementedException("'GetFormat' is not available in the 'CookieDateProvider'")
+            End Function
+        End Class
     End Module
 End Namespace