+++ SCrawler/My Project/AssemblyInfo.vb
@@ -32,6 +32,6 @@ Imports System.Runtime.InteropServices
 ' by using the '*' as shown below:
 ' <Assembly: AssemblyVersion("1.0.*")>
 
-<Assembly: AssemblyVersion("2026.3.23.0")>
-<Assembly: AssemblyFileVersion("2026.3.23.0")>
+<Assembly: AssemblyVersion("2026.5.13.0")>
+<Assembly: AssemblyFileVersion("2026.5.13.0")>
 <Assembly: NeutralResourcesLanguage("en")>